application-level adapter for server config and query string parsing
